# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Bokdy API
  version: 0.1.0
  description: |
    Bokdy foundation API (Identity + Organization + CRM + Catalog + Scheduling + Pricing
    + Reservation + Booking + Billing + Payment).
    Locale is taken from Accept-Language (missing → vi). Read models return a resolved name;
    owner edit payloads use name_en + name_vi.
    Auth: send X-Client (player|owner|admin) on register, login, refresh, logout.
    Timestamps are UTC RFC3339. Rate limit: 429 TOO_MANY_REQUESTS.
    Error envelope: { code, message, details? }. Codes are UPPERCASE; see packages/config/error-codes.json.
servers:
  - url: http://localhost:8088
    description: Local
...
paths:
...
  /api/v1/branches:
    get:
      tags: [Branches]
      summary: List branches
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XOrganizationId"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Branch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Branches]
      summary: Create branch
      description: Owner only. Initial status `inactive`. Creates empty location_settings (no scheduling rows).
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XOrganizationId"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Branch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/branches/{id}:
    get:
      tags: [Branches]
      summary: Get branch
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/XOrganizationId"
        - $ref: "#/components/parameters/AcceptLanguage"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Branch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Branches]
      summary: Update branch
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/XOrganizationId"
        - $ref: "#/components/parameters/AcceptLanguage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBranchRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Branch"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /api/v1/branches/{id}/open:
    post:
      tags: [Branches]
      summary: Open branch
      description: "inactive -> active"
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/XOrganizationId"
      responses:
        "204":
          description: Opened
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/branches/{id}/close:
    post:
      tags: [Branches]
      summary: Close branch
      description: "active -> inactive"
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/XOrganizationId"
      responses:
        "204":
          description: Closed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/branches/{id}/archive:
    post:
      tags: [Branches]
      summary: Archive branch
      description: "Any non-archived status -> archived. Booking checks deferred until Booking exists."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BranchId"
        - $ref: "#/components/parameters/XOrganizationId"
      responses:
        "204":
          description: Archived
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    BranchId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    XOrganizationId:
      name: X-Organization-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant organization context. Required for `/api/v1/branches*`, staff `/api/v1/customers*`, `/api/v1/court-types*`, `/api/v1/courts*`, staff schedule/availability routes, and staff booking routes (list, walk-in, confirm, check-in, complete).
...
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      schema:
        type: string
        default: vi
        example: vi
      description: BCP-47. First supported tag (vi, en) wins. Missing or unknown → vi.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Malformed JSON body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: BAD_REQUEST
            message: invalid json body
    ValidationFailed:
      description: Field validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: VALIDATION
            message: invalid request
            details:
              - field: email
                code: REQUIRED
                message: email is required
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - VALIDATION
            - BAD_REQUEST
            - TOO_MANY_REQUESTS
            - INTERNAL
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code:
          type: string
          enum:
            - REQUIRED
            - EMAIL_INVALID
            - TOO_SHORT
            - TOO_LONG
            - UUID_INVALID
            - ONE_OF
            - INVALID
            - PASSWORD_POLICY
        message: { type: string }
...
    BranchAddress:
      type: object
      required: [division_scheme]
      properties:
        division_scheme:
          $ref: "#/components/schemas/AdminDivisionScheme"
        country_id: { type: string, format: uuid }
        province_former_id: { type: string, format: uuid }
        district_former_id: { type: string, format: uuid }
        ward_former_id: { type: string, format: uuid }
        province_id: { type: string, format: uuid }
        ward_id: { type: string, format: uuid }
        address_line_1: { type: string }
        address_line_2: { type: string }
    AdminDivisionScheme:
      type: string
      enum: [former_v3, current_v2]
...
    Branch:
      type: object
      required: [id, public_id, organization_id, code, name, status]
      properties:
        id: { type: string, format: uuid }
        public_id: { type: string, minLength: 26, maxLength: 26 }
        organization_id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string, description: Resolved from Accept-Language }
        name_en: { type: string }
        name_vi: { type: string }
        phone: { type: string }
        email: { type: string, format: email }
        timezone: { type: string }
        status: { type: string, enum: [active, inactive, maintenance, archived] }
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/BranchAddress"
    CreateBranchRequest:
      type: object
      description: At least one of name_en or name_vi is required. Initial status is inactive.
      properties:
        name_en: { type: string }
        name_vi: { type: string }
        code: { type: string }
        phone: { type: string }
        email: { type: string, format: email }
        timezone: { type: string }
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/BranchAddress"
    UpdateBranchRequest:
      type: object
      properties:
        name_en: { type: string }
        name_vi: { type: string }
        code: { type: string }
        phone: { type: string }
        email: { type: string, format: email }
        timezone: { type: string }
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/BranchAddress"
...